fix: H5B nunca detectaba bajas de personal clave por un mismatch NUMERIC/FLOAT8
date - date entre fechas da integer en Postgres, y integer / 365.25 (literal numerico) promueve a NUMERIC, no FLOAT8. sqlx fallaba al decodificar como f64 (visible en logs de produccion) y get_f64 caia en su fallback silencioso a 0.0 -- eso alimentaba el filtro antiguedad_years >= 3.0, que nunca pasaba. H5B devolvia None siempre, para cualquier empresa.

Cast explicito a ::float8 en las dos consultas de compute_h5b (antiguedad activa y antiguedad al momento de la baja). Verificado contra CES100706U65 (el RFC del log): con el calculo corregido hay 7 bajas con 3+ anos de antiguedad real que antes nunca se mostraban.

+++ src/services/analytics/hallazgos.rs
@@ -812,7 +812,7 @@ async fn compute_h5b(pool: &DbPool, rfc: &str) -> anyhow::Result<Option<Hallazgo
         )
         SELECT
             lo.sueldo,
-            (((date_trunc('month', CURRENT_DATE) - interval '1 day')::date) - sd.start_date) / 365.25 AS antiguedad_years
+            ((((date_trunc('month', CURRENT_DATE) - interval '1 day')::date) - sd.start_date) / 365.25)::float8 AS antiguedad_years
         FROM last_ordinario lo
         JOIN start_dates sd ON sd.rfc_receptor = lo.rfc_receptor
         WHERE lo.sueldo > 0
@@ -910,7 +910,7 @@ async fn compute_h5b(pool: &DbPool, rfc: &str) -> anyhow::Result<Option<Hallazgo
             sd.start_date::text AS start_date,
             lr.fecha_final_pago::text AS fecha_final_pago,
             lo.sueldo,
-            (make_date((t.last_period/100)::int, (t.last_period%100)::int, 1) - sd.start_date) / 365.25 AS antiguedad_al_baja
+            ((make_date((t.last_period/100)::int, (t.last_period%100)::int, 1) - sd.start_date) / 365.25)::float8 AS antiguedad_al_baja
         FROM term t
         JOIN last_row lr ON lr.rfc_receptor = t.rfc_receptor
         JOIN last_ordinario lo ON lo.rfc_receptor = t.rfc_receptor
